v5: 目录重组 + 全库对拍体系 + 规则宪章三连
- 目录重组: 全引擎归位 algorithms/ 家族目录, 对拍测试集中各家族对拍/ 子目录
- 对拍体系: 12 套 *_check.cpp 全绿(conn/trie/mint/bigint/leftist/dsu/oset/pst/seg/rnd/vt/graph), 暴力独立构造
- utils 工具箱升格: 别名/最值常量/快读快写/debug/方向数组引用制, 全库禁内嵌副本
- rule.cpp v4: §3.5 高耦合引擎封装策略(LCT/SAM/网络流/莫队分档, 不用 CRTP), 对拍暴力哲学(独立构造+数学本质参照系), §3 设计目标校准(耦合场景人出错率最低)
- run_checks.ps1: 一键回归 12 套对拍

+++ algorithms/图论/欧拉图/例题/Luogu_P_1127.cpp
@@ -138,16 +138,16 @@ void solve()
         if(edges[i].size()) sort(edges[i].begin(), edges[i].end(), greater<PSI>());
     }
     dfs(s);
-    if(ans.size() != n)
+    if((int)ans.size() != n)
     {
         cout << "***" << endl;
         return;
     }
     reverse(ans.begin(), ans.end());
-    for(int i = 0; i < ans.size(); i++)
+    for(int i = 0; i < (int)ans.size(); i++)
     {
         cout << ans[i];
-        if(i != ans.size() - 1)
+        if(i != (int)ans.size() - 1)
         {
             cout << '.';
         }
